order_service/requirements.txt
fastapi
uvicorn
httpx
redis
python-jose
python-dotenv
sqlalchemy
asyncpg
